fix: widen SpscFifo drain index operands before addition (clang-tidy)
bugprone-misplaced-widening-cast: static_cast<size_t>(startN + i) widened
the int sum after the fact; widen each operand first so the arithmetic is
in size_t. Index is identical for all valid ranges (capacities 256/1024);
no FIFO logic changed.

Test Plan: 26/26 default ctest; 20/20 TSan (-L unit) incl. FIFO wrap +
concurrent stress; clang-tidy bugprone-misplaced-widening-cast resolved.

+++ engine/graph/SpscFifo.h
@@ -92,10 +92,10 @@ public:
         fifo.prepareToRead (ready, start1, size1, start2, size2);
 
         for (int i = 0; i < size1; ++i)
-            fn (std::as_const (storage[static_cast<std::size_t> (start1 + i)]));
+            fn (std::as_const (storage[static_cast<std::size_t> (start1) + static_cast<std::size_t> (i)]));
 
         for (int i = 0; i < size2; ++i)
-            fn (std::as_const (storage[static_cast<std::size_t> (start2 + i)]));
+            fn (std::as_const (storage[static_cast<std::size_t> (start2) + static_cast<std::size_t> (i)]));
 
         fifo.finishedRead (size1 + size2);
     }
